Consolidate entity validation in one module
